% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proLikelihood.R
\name{averageBE}
\alias{averageBE}
\title{Calculate profile likelihood for mean difference}
\usage{
averageBE(dat, colSpec = list(), theta = NULL, xlow, xup, xlength)
}
\arguments{
\item{dat}{data.frame}

\item{colSpec}{a named list that should specify columns in data; \sQuote{subject},
\sQuote{formula}, and \sQuote{y} are required. \sQuote{period} and \sQuote{seq}
may also be provided. The \sQuote{formula} column should identify treatment with R/T.}

\item{theta}{numeric vector see proLikelihood()}

\item{xlow}{numeric value see proLikelihood()}

\item{xup}{numeric value see proLikelihood()}

\item{xlength}{numeric value Defaults to 100; see proLikelihood()}
}
\value{
A \sQuote{proLikelihood} object, with elements \sQuote{poi},
\sQuote{maxLik}, \sQuote{MAX}, \sQuote{LI}, and \sQuote{method}. See proLikelihood().
}
\description{
This function is equivalent to proLikelihood() with method=\sQuote{average}.
}
\details{
See function proLikelihood()
}
\examples{
\donttest{
data(dat)
cols <- list(subject = 'subject', formula = 'formula', y = 'AUC')
l <- averageBE(dat, colSpec = cols, xlength = 300)
}

}
