% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beInternal}
\alias{beInternal}
\alias{design_matrix}
\alias{select_theta}
\alias{setup_env}
\alias{sigma_vals}
\alias{validateColumns}
\alias{varcov_matrix}
\title{Internal functions}
\description{
Internal functions are not intended to be called by user.
}
\details{
\code{design_matrix}: design matrix for given sequence

\code{select_theta}: select good starting values for theta

\code{setup_env}: create environment with required variables

\code{sigma_vals}: calculate sigma values used in varcov matrix

\code{validateColumns}: ensure accurate column specification

\code{varcov_matrix}: design varcov matrix
}
\keyword{internal}
