% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BElikelihood-package.R
\docType{package}
\name{BElikelihood-package}
\alias{BElikelihood}
\alias{BElikelihood-package}
\title{Likelihood method for analyzing bioequivalence (BE) trial data}
\description{
This package will calculate and plot the profile likelihoods for the mean difference and standard deviation ratios of a test drug 
to a reference drug for AUC or Cmax from various crossover designs commonly used in BE studies, such as a fully replicated crossover 
design (e.g., 2x4 two-sequence, four-period, RTRT/TRTR), a partially replicated crossover design (e.g., 2x3, two-sequence, three-period, RTR/TRT), and a two-sequence, two-period, crossover design design (2x2, RT/TR), where "R" stands for a reference drug and "T" stands for a test drug.
}
\author{
\strong{Maintainer}: Liping Du \email{liping.du@vumc.org} (\href{https://orcid.org/0000-0003-4926-2135}{ORCID})

Authors:
\itemize{
  \item Leena Choi \email{naturechoi@gmail.com} (\href{https://orcid.org/0000-0002-2544-7090}{ORCID})
  \item Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})
}

}
