% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printplot.R
\name{plot.proLikelihood}
\alias{plot.proLikelihood}
\title{Plot method for proLikelihood object}
\usage{
\method{plot}{proLikelihood}(x, textx, texty = 0.9, textsize = 3, ...)
}
\arguments{
\item{x}{proLikelihood object}

\item{textx}{numeric value position of label}

\item{texty}{numeric value position of label}

\item{textsize}{numeric value text size of label}

\item{\dots}{unused}
}
\value{
ggplot2 object
}
\description{
function summary
}
\details{
function details
}
\examples{
\donttest{
data(dat)
cols <- list(subject = 'subject', formula = 'formula', y = 'AUC')
p4a <- proLikelihood(dat, colSpec = cols, xlength = 300, method = 'average')
p4t <- proLikelihood(dat, colSpec = cols, xlength = 300, method = 'total')
p4w <- proLikelihood(dat, colSpec = cols, xlength = 300, method = 'within')
plot(p4a)
plot(p4t)
plot(p4w)
# three period case
dd3 <- dat[dat$period < 4,]
p3a <- proLikelihood(dd3, colSpec = cols, xlength = 300, method = 'average')
plot(p3a)
# two period case
dd2 <- dat[dat$period < 3,]
p2a <- proLikelihood(dd2, colSpec = cols, xlength = 300, method = 'average')
plot(p2a)
}

}
