% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printplot.R
\name{print.proLikelihood}
\alias{print.proLikelihood}
\title{Print method for proLikelihood object}
\usage{
\method{print}{proLikelihood}(x, ...)
}
\arguments{
\item{x}{proLikelihood object}

\item{\dots}{unused}
}
\value{
return \sQuote{x} invisibly
}
\description{
Print \sQuote{poi} and \sQuote{maxLik} elements from a proLikelihood object.
}
\examples{
\donttest{
data(dat)
cols <- list(subject = 'subject', formula = 'formula', y = 'AUC')
l <- averageBE(dat, colSpec = cols, xlength = 300)
l
}

}
