% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proLikelihood.R
\name{proLikelihood}
\alias{proLikelihood}
\title{Run profile likelihood}
\usage{
proLikelihood(
  dat,
  colSpec = list(),
  theta = NULL,
  xlow,
  xup,
  xlength = 100,
  method
)
}
\arguments{
\item{dat}{data.frame}

\item{colSpec}{a named list that should specify columns in data; \sQuote{subject},
\sQuote{formula}, and \sQuote{y} are required. \sQuote{period} and \sQuote{seq}
may also be provided. The \sQuote{formula} column should identify treatment with R/T.}

\item{theta}{numeric vector DESCRIBE}

\item{xlow}{numeric value DESCRIBE}

\item{xup}{numeric value DESCRIBE}

\item{xlength}{numeric value Defaults to 100; DESCRIBE}

\item{method}{character value Should be one of \sQuote{average}, \sQuote{total},
or \sQuote{within}. DESCRIBE}
}
\value{
A \sQuote{proLikelihood} object, with elements \sQuote{poi},
\sQuote{maxLik}, \sQuote{MAX}, \sQuote{LI}, and \sQuote{method}.
}
\description{
function summary
}
\details{
function details
}
\examples{
\donttest{
data(dat)
cols <- list(subject = 'subject', formula = 'formula', y = 'AUC')
p4a <- proLikelihood(dat, colSpec = cols, xlength = 300, method = 'average')
p4t <- proLikelihood(dat, colSpec = cols, xlength = 300, method = 'total')
p4w <- proLikelihood(dat, colSpec = cols, xlength = 300, method = 'within')
# three period case
dd3 <- dat[dat$period < 4,]
p3a <- proLikelihood(dd3, colSpec = cols, xlength = 300, method = 'average')
p3t <- proLikelihood(dd3, colSpec = cols, xlength = 300, method = 'total')
p3w <- proLikelihood(dd3, colSpec = cols, xlength = 300, method = 'within')
# two period case
dd2 <- dat[dat$period < 3,]
p2a <- proLikelihood(dd2, colSpec = cols, xlength = 300, method = 'average')
p2t <- proLikelihood(dd2, colSpec = cols, xlength = 300, method = 'total')
}

}
