% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proLikelihood.R
\name{totalVarianceBE}
\alias{totalVarianceBE}
\title{Run profile likelihood}
\usage{
totalVarianceBE(dat, colSpec = list(), theta = NULL, xlow, xup, xlength)
}
\arguments{
\item{dat}{data.frame}

\item{colSpec}{a named list that should specify columns in data; \sQuote{subject},
\sQuote{formula}, and \sQuote{y} are required. \sQuote{period} and \sQuote{seq}
may also be provided. The \sQuote{formula} column should identify treatment with R/T.}

\item{theta}{numeric vector DESCRIBE}

\item{xlow}{numeric value DESCRIBE}

\item{xup}{numeric value DESCRIBE}

\item{xlength}{numeric value Defaults to 100; DESCRIBE}
}
\value{
A \sQuote{proLikelihood} object, with elements \sQuote{poi},
\sQuote{maxLik}, \sQuote{MAX}, \sQuote{LI}, and \sQuote{method}.
}
\description{
function summary
}
\details{
function details
}
\examples{
\donttest{
data(dat)
cols <- list(subject = 'subject', formula = 'formula', y = 'AUC')
tv <- totalVarianceBE(dat, colSpec = cols, xlength = 300)
}

}
