% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proLikelihood.R
\name{withinVarianceBE}
\alias{withinVarianceBE}
\title{Calculate profile likelihood for within-subject standard deviation ratio of test to reference drug}
\usage{
withinVarianceBE(dat, colSpec = list(), theta = NULL, xlow, xup, xlength)
}
\arguments{
\item{dat}{data.frame}

\item{colSpec}{a named list that should specify columns in data; \sQuote{subject}, 
\sQuote{formula}, and \sQuote{y} are required. \sQuote{period} and \sQuote{seq} 
may also be provided. The \sQuote{formula} 
column should identify either the reference and test drug using R and T, respectively, where other values are not allowed.}

\item{theta}{numeric vector. See proLikelihood()}

\item{xlow}{numeric value. See proLikelihood()}

\item{xup}{numeric value. See proLikelihood()}

\item{xlength}{numeric value Defaults to 100. See proLikelihood()}
}
\value{
A \sQuote{proLikelihood} object, with elements \sQuote{poi},
\sQuote{maxLik}, \sQuote{MAX}, \sQuote{LI}, and \sQuote{method}. See proLikelihood()
}
\description{
This function is equivalent to proLikelihood() with method=\sQuote{within}.
}
\details{
see proLikelihood()
}
\examples{
\donttest{
data(dat)
cols <- list(subject = 'subject', formula = 'formula', y = 'AUC')
wv <- withinVarianceBE(dat, colSpec = cols, xlength = 300)
}

}
