% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BElikelihood-package.R
\docType{package}
\name{BElikelihood-package}
\alias{BElikelihood}
\alias{BElikelihood-package}
\title{Bioequivalence Likelihood}
\description{
need description
}
\author{
\strong{Maintainer}: Liping Du \email{liping.du@vumc.org}

Authors:
\itemize{
  \item Leena Choi \email{leena.choi@vanderbilt.edu} (\href{https://orcid.org/0000-0002-2544-7090}{ORCID})
  \item Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})
}

}
