% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BElikelihood-package.R
\docType{package}
\name{BElikelihood-package}
\alias{BElikelihood}
\alias{BElikelihood-package}
\title{Likelihood method for analyzing Bioequivalence}
\description{
This package will calculate and plot the profile likelihoods for the mean difference and standard deviation ratios of a test drug to a reference drug for
AUC or Cmax from a full replicate 2x4 (RTRT/TRTR) cross-over, 2x3 partial replicate (TRT/RTR) cross-over, or a 2x2 (RT/TR) cross-over design (where R and T for reference and test drugs, respectively).
}
\author{
\strong{Maintainer}: Liping Du \email{liping.du@vumc.org}

Authors:
\itemize{
  \item Leena Choi \email{leena.choi@vanderbilt.edu} (\href{https://orcid.org/0000-0002-2544-7090}{ORCID})
  \item Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})
}

}
