% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{dat}
\format{
A data frame with 176 observations on 6 variables.
}
\usage{
data(dat, package = 'BElikelihood')
}
\description{
Describe example data.
}
\examples{
data(dat)
}
\keyword{datasets}
